--- src/main/java/com/example/voucher/DiscountPolicy.java
@@ -1,5 +1,0 @@
-package com.example.voucher;
-
-public interface DiscountPolicy {
-    void applyDiscount(Voucher voucher);
-}
