+++ src/main/java/com/example/voucher/domain/discount/DiscountPolicy.java
@@ -1,0 +1,7 @@
+package com.example.voucher.domain.discount;
+
+import com.example.voucher.domain.Voucher;
+
+public interface DiscountPolicy {
+    void applyDiscount(Voucher voucher);
+}
